pyproject.toml
[tool.poetry]
name = "core-django"
version = "0.1.0"
description = ""
authors = ["Tony <tonyf132@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "core_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
pillow = "^10.0.0"
python-decouple = "^3.8"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
